vim: add mason and mason-lspconfig
I want to try out codelldb for Rust, and it appears as though mason is
the only reasonable mechanism I can find to programmatically install it.

Migrate my setup over to mason.  The only current change is to use
mason-lspconfig to automatically install pyright (as opposed to the
system's package manager).